+++ package.json
@@ -1,4 +1,3 @@
-
 {
   "name": "nile-cash",
   "version": "0.1.0",
@@ -12,6 +11,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-dotenv": "^0.1.3",
+    "react-firebase-hooks": "^5.1.1",
     "react-router-dom": "^6.21.3",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
