+++ src/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Navbar from './components/Navbar';
-import History from './components/History';
-import Homepage from './components/Homepage';
+import History from './pages/LoanHistory';
+import Homepage from './pages/Homepage';
 import './styles/App.css';
-import Form from './components/Form';
+import LoanRequestForm from './forms/RequestLoanForm';
 
 const App = () => {
   return (
@@ -12,9 +12,9 @@ const App = () => {
       <div className="app">
         <Navbar />
         <Routes>
-          <Route path="/history" element={<History />} />
-          <Route path="/loan_request" element={<Form />} />
           <Route path="/" element={<Homepage />} />
+          <Route path="/history" element={<History />} />
+          <Route path="/loan-request" element={<LoanRequestForm />} />
         </Routes>
 
       </div>
