+++ src/components/HeroSection.js
@@ -11,7 +11,7 @@ const HeroSection = () => {
       <div className='hero-left'>
         <h1>Get quick cash as an EIT!</h1>
         <p>Click request loan below to fill the request form. Our team will verify your request and dispense immediately. No collateral, no hidden charges.</p>
-        <button className="request-loan-button">Request Loan</button>
+        <button className="request-loan-button">Request Loan</button> 
         <div className="active-users">
           <span>51+ Active users at Nile house Accra, Ogbojo</span>
         </div>
@@ -20,8 +20,6 @@ const HeroSection = () => {
       <div className="hero-right">
         <img src={people} alt='people'></img>
       </div>
-
-
 
     </section>
   );
